Remove pytest-bdd
