+++ tests/functional/test_activemethod.py
@@ -1,22 +1,15 @@
 from dataclay import api
 import pytest
 from model.family import Family, Person, Dog
+from utils import init_client, mock_env_client
 
 
-@pytest.fixture
-def mock_env_client(monkeypatch):
-    monkeypatch.setenv("METADATA_SERVICE_HOST", "127.0.0.1")
-    monkeypatch.setenv("DC_USERNAME", "user")
-    monkeypatch.setenv("DC_PASSWORD", "s3cret")
-    monkeypatch.setenv("DEFAULT_DATASET", "myDataset")
-
-
-def test_activemethod_argument_make_persistent(mock_env_client):
+def test_activemethod_argument_make_persistent(init_client):
     """
     A dataclay object is made persistent when passed as an argument
     to an activemethod of a persistent object
     """
-    api.init()
+    # api.init()
 
     family = Family()
     person = Person("Marc", 24)
@@ -28,11 +21,11 @@ def test_activemethod_argument_make_persistent(mock_env_client):
     assert person == family.members[0]
 
 
-def test_activemethod_persistent_argument(mock_env_client):
+def test_activemethod_persistent_argument(init_client):
     """
     Persistent objects can be sent as activemethod arguments
     """
-    api.init()
+    # api.init()
     family = Family()
     person = Person("Marc", 24)
     family.make_persistent()
@@ -41,11 +34,11 @@ def test_activemethod_persistent_argument(mock_env_client):
     assert person == family.members[0]
 
 
-def test_activemethod_defined_properties(mock_env_client):
+def test_activemethod_defined_properties(init_client):
     """
     Object properties defined in class annotations are sychronized between the client and backend
     """
-    api.init()
+    # api.init()
     person = Person("Marc", 24)
     assert person.age == 24
 
@@ -54,11 +47,11 @@ def test_activemethod_defined_properties(mock_env_client):
     assert person.age == 25
 
 
-def test_activemethod_non_defined_properties(mock_env_client):
+def test_activemethod_non_defined_properties(init_client):
     """
     Object properties not defined in class annotations are not synchronized between the client and backend
     """
-    api.init()
+    # api.init()
     dog = Dog("Duna", 6)
     assert dog.age == 6
